Исправление критический замечаний ревьюера. Улучшение по замечаниям
