+++ README.md
@@ -1,12 +1,24 @@
 # Sprint_7_project_scooter
+## Тестирование API сервиса Яндекс.Самокат
+Документация по API - https://qa-scooter.praktikum-services.ru/docs/
+###
+| **Использованные библиотеки** |
+|-------------------------------|
+| java 11                       |
+| junit 4.13.2                  |
+| maven 3.9.6                   |
+| RestAssured 5.3.0              |
+| gson 2.10.1                   |
+| maven-surefire 2.22.2         |
+| allure-maven 2.15.0           |
 
-# Запуск тестов
+#### Запуск тестов
 ```
 mvn clean test
 ```
 
 
-# Создание отчета Allure
+##### Создание отчета Allure
 ```
 mvn allure:serve
 ```
