+++ src/main/java/courier/CourierAsserts.java
@@ -11,18 +11,21 @@ public class CourierAsserts {
     @Step("Регистрация курьера с валидными данными")
     public void courierCreate(ValidatableResponse response) {
         response.assertThat()
+                .statusCode(201)
                 .body("ok", is(true));
     }
 
     @Step("Ответ если не хватает данных")
     public void courierCreateError(ValidatableResponse response) {
         response.assertThat()
+                .statusCode(400)
                 .body("message", equalTo("Недостаточно данных для создания учетной записи"));
     }
 
     @Step("Ответ если пользователь уже существует")
     public void courierCreateLoginDuplicat(ValidatableResponse response) {
         response.assertThat()
+                .statusCode(409)
                 .body("message", equalTo("Этот логин уже используется. Попробуйте другой."));
     }
 
@@ -37,12 +40,19 @@ public class CourierAsserts {
     @Step("Ответ при нехватки данных для входа")
     public void courierLoginWithoutData(ValidatableResponse response) {
         response.assertThat()
+                .statusCode(400)
                 .body("message", equalTo("Недостаточно данных для входа"));
     }
 
     @Step("Ответ при невалидных данных авторизации")
     public void courierLoginWithIncorrectData(ValidatableResponse response) {
         response.assertThat()
+                .statusCode(404)
                 .body("message", equalTo("Учетная запись не найдена"));
+    }
+    @Step("Успешный вход пользователя")
+    public void courierValidLogin(ValidatableResponse response){
+        response.assertThat()
+                .statusCode(200);
     }
 }
